requirements.txt
# Razer Nari Pairing Tool - Requirements
pyusb>=1.2.1
